Simplified image build, created fleet provisioning script
This commit moves the project towards creating a smaller golden image
with no node-specific logic. All instance provisioning now happens in
configure-fleet.sh, a script that takes over the functionality of
node-specific scripts in the scripts/ directory and node-specific
logic in the Packer build file. This allows for an operator to
provision an entire fleet from a single machine in a single command
instead of hopping between each node and the build machine for no
clear reason.