Rewrite it in Rust
This commit introduces a more or less complete rewrite of the
provisioning workflow as a DAG in Rust.

At a high level, the Graph struct is made up of nodes and edges
between them. A graph walk to a target node works by computing the
subgraph of all source nodes that can reach the target, then awaiting
each node's source nodes before running the given node's visit
function. This allows for independent steps (e.g., NFS and iSCSI
mounts) to run concurrently both within and between individual
instances.

Remaining work is mostly in providing better names for things;
removing sleeps in favor of checking for the existence of required
devices files; and documenting various structs, functions, and so on.